requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
SQLAlchemy==2.0.23
pandas==2.1.4
openpyxl==3.1.5
WTForms==3.1.1
Flask-WTF==1.2.1
python-dotenv==1.0.0
APScheduler==3.10.4
Flask-Mail==0.9.1
Werkzeug==3.0.1
